whole whitted algorithm, compiles, but draw nothing. Function trace in cast_ray never returns true. debugging needed

+++ libft/ft_swap.c
@@ -12,11 +12,20 @@
 
 #include "libft.h"
 
-void	ft_swap(int *a, int *b)
+void	ft_swap(double *a, double *b)
 {
-	int temp;
+	double temp;
 
 	temp = *a;
 	*a = *b;
 	*b = temp;
 }
+
+//void	ft_swap(int *a, int *b)
+//{
+//	int temp;
+//
+//	temp = *a;
+//	*a = *b;
+//	*b = temp;
+//}
