+++ libft/libft.h
@@ -85,7 +85,7 @@ void				ft_lstadd(t_list **alst, t_list *new);
 void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list				*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
 char				*ft_strrev(char *str);
-void				ft_swap(int *a, int *b);
+void				ft_swap(double *a, double *b);
 char				*ft_uppercase(char *str);
 char				*ft_lowercase(char *str);
 size_t				ft_strlcpy(char *dst, const char *src, size_t size);
